check if folder exists first

+++ Tab/AudioHandler.cs
@@ -61,7 +61,8 @@ internal class AudioHandler
         //  var fileNameB = "b.webm";
         var remoteIpProxiedHeader = context.Request.Headers["CF-Connecting-IP"].FirstOrDefault() ?? context.Connection.RemoteIpAddress.ToString();
         var tempFolder = Path.Combine(Path.GetTempPath(), "audio" + remoteIpProxiedHeader);
-        Directory.Delete(tempFolder, true);
+        if (Directory.Exists(tempFolder))
+            Directory.Delete(tempFolder, true);
         Directory.CreateDirectory(tempFolder);
         var names = new NameHandler(tempFolder);
         var fileName = names.GetCurrent();
